requirements.txt
Django==4.2.4
isort==5.12.0
black==23.7.0
flake8==6.1.0
djangorestframework==3.14.0
markdown==3.4.4
django-filter==23.2
djangorestframework-simplejwt==5.3.0